Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -14,7 +14,7 @@ int main(){
     else if (c=='*'){
         printf("%d",a*b);
     }
-    double Quotient==a/b;
+    float Quotient==a/b;
     else if (c=='/'){
         printf("%d",a/b,Quotient);
     }
